Add calculator class
