+++ types.h
@@ -6,12 +6,4 @@ struct EXP_DIVIDER_RESULT {
     int right_value;
 };
 
-struct PROGRESSION_FLAGS {
-    list<string> expressions;
-    list<string>::iterator it;
-    bool exists_mul_and_div;
-    bool exists_pls_and_mns;
-    bool go_through;
-};
-
 #endif
